# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SIH26146 — Bitcoin Risk Analysis API
  description: AI-powered monitoring and analysis of Bitcoin transaction traffic.
    Pipeline B (Bitcoin GCN) produces model-derived risk rankings (0–100) for address
    prioritization. Scores are NOT calibrated probabilities and do not constitute
    an enforcement or legal determination. Every response that contains a risk score
    includes a score_disclaimer field.
  version: 0.1.0
paths:
...
  /api/v1/wallets/{address_hash}:
    get:
      tags:
      - wallets
      summary: Full risk profile for a single address
      description: Returns model-derived risk score (0–100), risk label, up to 3 explanation
        reasons, graph-derived feature values, co-spend cluster membership, and P2P
        network signals (always null in demo mode). score_disclaimer is always present
        and must be displayed in the UI.
      operationId: get_wallet_api_v1_wallets__address_hash__get
      parameters:
      - name: address_hash
        in: path
        required: true
        schema:
          type: string
          title: Address Hash
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    P2PSignal:
      properties:
        signal_source:
          type: string
          enum:
          - SIMULATED
          const: SIMULATED
          title: Signal Source
          default: SIMULATED
        ip_cluster_id:
          type: string
          title: Ip Cluster Id
          examples:
          - IP_CLUSTER_07
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        tx_hash:
          type: string
          title: Tx Hash
          examples:
          - a0b05f87c91a36014d07...
        disclaimer:
          type: string
          enum:
          - Simulated/replayed demonstration data only.
          const: Simulated/replayed demonstration data only.
          title: Disclaimer
          default: Simulated/replayed demonstration data only.
      type: object
      required:
      - ip_cluster_id
      - timestamp
      - tx_hash
      title: P2PSignal
      description: 'Simulated P2P network signal (P0.7).

        signal_source is always ''SIMULATED'' — this is a demonstration artefact.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WalletDetailResponse:
      properties:
        score_disclaimer:
          type: string
          title: Score Disclaimer
          description: Hard-coded disclaimer that must accompany every risk score.
            NOT a calibrated probability. For human review and prioritization only.
          default: Model-derived risk ranking. For prioritization and human review
            only. Not a calibrated probability.
          examples:
          - Model-derived risk ranking. For prioritization and human review only.
            Not a calibrated probability.
        address:
          type: string
          title: Address
        risk_score:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Risk Score
        risk_label:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - CRITICAL
          title: Risk Label
        reasons:
          items:
            type: string
          type: array
          maxItems: 3
          title: Reasons
        hops_to_nearest_flagged:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hops To Nearest Flagged
          description: Shortest-path distance to nearest known-flagged cluster. null
            when no flagged cluster is reachable within the snapshot.
        tx_burst_score:
          type: number
          title: Tx Burst Score
          description: Feature F11 from FEATURE_SPEC.md — ratio of tx count to active
            hours.
        timing_anomaly_score:
          type: number
          title: Timing Anomaly Score
          description: Feature F12 — weekday vs weekend transaction ratio.
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        cluster_confidence:
          anyOf:
          - type: string
            enum:
            - HIGH
            - MEDIUM
            - LOW
          - type: 'null'
          title: Cluster Confidence
        candidate_change_address:
          type: boolean
          title: Candidate Change Address
          description: True if this address was identified as a likely change-address
            by the P0.4.3 heuristic (single-input, two-output, fresh address, smaller
            output value). This is a heuristic inference, not a certainty.
        p2p_signals:
          anyOf:
          - items:
              $ref: '#/components/schemas/P2PSignal'
            type: array
          - type: 'null'
          title: P2P Signals
          description: Simulated P2P signals. Always null until P0.7 is implemented.
      type: object
      required:
      - address
      - risk_score
      - risk_label
      - reasons
      - tx_burst_score
      - timing_anomaly_score
      - candidate_change_address
      title: WalletDetailResponse
      description: 'GET /api/v1/wallets/{address_hash}


        Full risk profile for a single address.'